Fixes: #1601
changed the way the ROI painter draws the disk brush, to match she shape of the ball one. It fixes the painter / cursor mismatch, which was actually due to int rounding mistakes. For small sizes (2), the cursor/brush have more or less the shape of a cross now, which is thinner than the 3x3 square as before. But we cannot build a 2x2 cursor because of the symmetry around the central voxel, which is a bit a pity since such a brush whould be useful. Maybe by un-centering the brush from the center of the voxel...



git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/branches/4.0@38910 613412ce-1361-4b7a-93bf-a3c75af1bae4