Do not assume that identically named referentials are the same
Referentials that are referenced by name instead of UUID should be
considered different unless the user wants to merge them. A common
example is the NIFTI_XFORM_ALIGNED_ANAT referential (a.k.a.
"Coordinates aligned to another file or to anatomical truth"),
which is a very generic description and in practice never refers
to the same actual referential.