changed "symmetric" (zero-centered) palettes behavior:
max is now an absolute (or reversed) max, and min sets the threshold in
the center of the palette. This way we can threshold positive/negative
values using the min.
All palette and texture image code had to be reworked. ColorTraits is
now the central and mandatory way to use color palettes.