don't register objects in the load notification event
objects are now registered as soon as they are loaded, even from
different threads, now that registration is thread-safe.
It used to be done in an event, which run after other code in the main
thread, thus object references could have already been modified, thus
reference counting could be messed-up.
(this only happened in the new LoadObjectsCommand, however)