workaround fix for widgets inside graphics views
Since we have switched to QOpenGLWidget, graphics views with GAGMWidget
as a viewport don't display texts.
After struggling with it I found out that widgets inserted as windows,
not frameless, with opacity != 1, and with a window title, do work (!)
I could not figure out a better solution, so now we insert widgets with
window borders and title bar.
It seems to be a repaint/update problem, but I can't find out what
triggers it. Using a "real" grapicsview and QOpenGLWidget do not have
this problem.